RemoteConfigUtil added
